# src/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: Royal Burger API
  description: API para o sistema de delivery da hamburgueria Royal Burger.
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000/api
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    MessageResponse:
      type: object
      properties: {msg: {type: string, example: "Operação realizada com sucesso."}}
...
    Product:
      type: object
      properties: {id: {type: integer, example: 1}, name: {type: string, example: "X-Salada Clássico"}, description: {type: string, example: "Um clássico pão, hambúrguer de 150g, queijo mussarela, alface e tomate frescos."}, price: {type: number, format: float, example: 19.90}, image_url: {type: string, example: "https://seusite.com/imagens/x-salada.png"}, section_id: {type: integer, example: 2}, is_active: {type: boolean, example: true}}
    NewProduct:
      type: object
      required: [name, price, section_id]
      properties: {name: {type: string, example: "Batata Frita Especial"}, description: {type: string, example: "Porção de 300g de batatas fritas crocantes com cheddar e bacon."}, price: {type: number, format: float, example: 15.00}, image_url: {type: string, example: "https://seusite.com/imagens/batata-frita.png"}, section_id: {type: integer, example: 3}}
    ProductIngredient:
      type: object
      properties:
        ingredient_id: {type: integer, example: 15}
        name: {type: string, example: "Picles"}
        quantity: {type: number, format: float, example: 4.0, description: "A quantidade do ingrediente na receita (ex: 2.0 para duas fatias)"}
    NewProductIngredient:
      type: object
      required: [ingredient_id, quantity]
      properties: {ingredient_id: {type: integer, example: 10}, quantity: {type: number, format: float, example: 2.0}}
...
    User:
      type: object
      properties: {id: {type: integer, example: 2}, full_name: {type: string, example: "Gerente Silva"}, email: {type: string, format: email, example: "gerente@royalburger.com"}, role: {type: string, enum: [admin, manager, attendant], example: "manager"}, is_active: {type: boolean, example: true}}
    NewUser:
      type: object
      required: [full_name, email, password, role]
      properties: {full_name: {type: string, example: "Atendente Carlos"}, email: {type: string, format: email, example: "carlos@royalburger.com"}, password: {type: string, example: "SenhaSegura456"}, role: {type: string, enum: [admin, manager, attendant], example: "attendant"}}
    UserLogin:
      type: object
      required: [email, password]
      properties: {email: {type: string, example: "admin@royalburger.com"}, password: {type: string, example: "supersecret"}}
    TokenResponse:
      type: object
      properties: {access_token: {type: string, example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}
    PasswordResetRequest:
      type: object
      required: [email]
      properties: {email: {type: string, format: email, example: "gerente@royalburger.com"}}
    PasswordResetFinalize:
      type: object
      required: [token, new_password]
      properties: {token: {type: string, example: "um_token_jwt_especial_de_reset"}, new_password: {type: string, example: "NovaSenha@123"}}
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      summary: Lista todos os produtos ativos (Cardápio).
      security: []
      tags: [Produtos]
      responses:
        '200':
          description: Cardápio retornado com sucesso.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Product'}
    post:
      summary: (Admin) Cria um novo produto.
      tags: [Produtos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '201':
          description: Produto criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{product_id}:
    get:
      summary: Busca os detalhes de um produto específico.
      security: []
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes do produto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      summary: (Admin) Atualiza um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProduct'
      responses:
        '200':
          description: Produto atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      summary: (Admin) Inativa um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto inativado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /products/{product_id}/ingredients:
    get:
      summary: Lista os ingredientes de um produto.
      security: []
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de ingredientes do produto.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProductIngredient'}
    post:
      summary: (Admin) Adiciona um ingrediente a um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProductIngredient'
      responses:
        '201':
          description: Ingrediente associado/atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /products/{product_id}/ingredients/{ingredient_id}:
    delete:
      summary: (Admin) Remove um ingrediente de um produto.
      tags: [Produtos]
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
        - name: ingredient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente desassociado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /sections/{section_id}/products/{product_id}:
    post:
      summary: (Admin) Associa um produto a uma seção.
      tags: [Seções]
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Associação realizada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      summary: (Admin) Remove a associação de um produto a uma seção.
      tags: [Seções]
      parameters:
        - name: section_id
          in: path
          required: true
          schema:
            type: integer
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Associação removida com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /users/login:
    post:
      summary: Autentica um funcionário.
      security: []
      tags: [Autenticação (Funcionários)]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Autenticação bem-sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /users/request-password-reset:
    post:
      summary: Inicia o processo de redefinição de senha.
      security: []
      tags: [Autenticação (Funcionários)]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Se o e-mail existir, uma instrução será enviada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /users/reset-password:
    post:
      summary: Finaliza a redefinição de senha.
      security: []
      tags: [Autenticação (Funcionários)]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetFinalize'
      responses:
        '200':
          description: Senha redefinida com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /users:
    get:
      summary: (Admin) Lista todos os funcionários.
      tags: [Usuários (Funcionários)]
      responses:
        '200':
          description: Lista de funcionários.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      summary: (Admin) Cria um novo funcionário.
      tags: [Usuários (Funcionários)]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '201':
          description: Funcionário criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    get:
      summary: (Admin) Busca um funcionário por ID.
      tags: [Usuários (Funcionários)]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dados do funcionário.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: (Admin) Atualiza um funcionário.
      tags: [Usuários (Funcionários)]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
      responses:
        '200':
          description: Funcionário atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
    delete:
      summary: (Admin) Inativa um funcionário.
      tags: [Usuários (Funcionários)]
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Funcionário inativado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'